Make XML processing deterministic
Prior to this patch, each build of an identical Hostboot tree
would result in differing obj/genfiles - which since many of
these are included *everywhere*, force the entirety of hostboot
to be rebuilt when using ccache rather than hitting the cache.

It also means that repeated builds are likely to end up with
differing binaries, thwarting efforts for repeatable builds.

With this patch, rebuilding hostboot as part of op-build goes from
around 16 minutes to less than 3 minutes.

What this patch does is modifies several of the Perl scripts used
to parse XML and other input so that their output is deterministic.

In many places, these scripts did a "foreach my $key (%hash)" which
will iterate over the hash in *hash value* order, which is not
deterministic from one machine to the next, nor one invocation to
the next.

I modify these locations (based on diffing obj/genfiles and looking
at what differed) to iterate based on something that's orderable
(such as attribute name). This has the added side effect of making
the output a bit nicer to look at too.

I *think* I got all of these locations correct, and my testing
has shown that it appears to produce a binary out the end.

You will need:
 - an op-build patch as well that will ensure that the hostboot
   build is invoked with CCACHE (previously, it was not)
   (hopefully merged by the time you read this)
 - BR2_CCACHE=y in your configuration (if using op-build)
 - A ccache size adequate for Hostboot.
   NOTE A single hostboot build is 7.1GB of ccache, the default
        ccache size is 5GB.
   i.e. add BR2_CCACHE_INITIAL_SETUP="--max-size 20G" to your config

Change-Id: I7dac66a30be0a2b4c10f308211bedd30f57b851c
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>