XZ compressed HBI
This shaves about 6 seconds off the time spent in hostboot on a
Witherspoon system by alleviating PNOR IO over the (rather slow) LPC
bus. Nearly all of this time (5 seconds) is saved while cache contained,
where memory pressure exists and is relatively high.

For Secure Boot to work with on-demand paging, there exists a table of
cryptographic hashes for each page that can be brought in. The secure
boot code verifies this table using the normal mechanism, and then, when
each page is paged in, the content of the read page is verified against
the stored hash.

The new HBI format with this patch includes the (now modified) table of
hashes, and after that a concatenation of each compressed HBI page. This
takes the total size of the HBI partition down from ~15MB to ~5MB.

What this patch does is extend that hash page table (not to be confused
with the MMU mode) to include enough metadata to find the compressed
page. Since a compressed page is naturally smaller than an uncompressed
one, we can no longer do simple multiplication to work out where a page
is.

The downside is increased memory usage for the needed buffers for XZ
decompression, as well as more CPU usage for doing the decompression.
Luckily, since LPC is a whopping 1.5MB/sec, we're probably always going
to win on CPU versus IO tradeoff.

Change-Id: I23895fcb1e2193d4b359986ca56a5435b327ce91
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>